[REFACTOR] constants 경로 추가
